tjmooc/requirements.txt
Django==1.10.5
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
drfdocs==0.0.11
jsonpickle==0.9.4
olefile==0.44
Pillow==4.0.0
PyJWT==1.4.2
